requirements.txt
python-telegram-bot==20.7
PyGithub==2.1.1
python-dotenv==1.0.0
Pillow==10.2.0


